Type decoded JWT, move AuthenticatedUser to types file
