+++ frontend/src/APIClients/BaseAPIClient.ts
@@ -5,6 +5,8 @@ import jwt from "jsonwebtoken";
 import AUTHENTICATED_USER_KEY from "../constants/AuthConstants";
 import { setLocalStorageObjProperty } from "../utils/LocalStorageUtils";
 
+import { DecodedJWT } from "../types/AuthTypes";
+
 const baseAPIClient = axios.create({
   baseURL: process.env.REACT_APP_BACKEND_URL,
 });
@@ -33,12 +35,12 @@ baseAPIClient.interceptors.request.use(async (config: AxiosRequestConfig) => {
     authHeaderParts.length >= 2 &&
     authHeaderParts[0].toLowerCase() === "bearer"
   ) {
-    // TODO: define expected JWT return type
-    const decodedToken: any = jwt.decode(authHeaderParts[1]);
+    const decodedToken = jwt.decode(authHeaderParts[1]) as DecodedJWT;
 
     if (
       decodedToken &&
-      decodedToken.exp <= Math.round(new Date().getTime() / 1000)
+      (typeof decodedToken === "string" ||
+        decodedToken.exp <= Math.round(new Date().getTime() / 1000))
     ) {
       const { data } = await axios.post(
         `${process.env.REACT_APP_BACKEND_URL}/auth/refresh`,
